% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2_helper_functions.R
\name{AMR_build}
\alias{AMR_build}
\title{Build AMR vector}
\usage{
AMR_build(res, methylation, nb_cpg = 2)
}
\arguments{
\item{res}{result object of function AMR_search}

\item{methylation}{a matrix of methylation profile.}

\item{nb_cpg}{threshold of minimal number of CpG in the AMR}
}
\value{
A set of build AMRs.
\itemize{
\item res, selected AMR
\item CpG_for_each_AMR, list of markers present on each AMR.
\item AMR_acp, first components of PCA for each DMR
\item AMR_mean, mean value of CpG on the AMR
}
}
\description{
Build AMR from the result of function AMR_search
}
\details{
We use the series of pValues (one pValue per CpGs) obtained with the mEWAS
regression method and the combination of pValue max2.
To determine the potential AMRs used the combp method present in the ENmix package (Xu et al. 2016).
This method uses the Fisher method to combine the pValues and also the base pair distance (bP)
between CpGs (1000 bP maximum between nb_cpg CpGs on the same AMR).
The information for each AMR is summarized by doing the mean (by row) of each CpG.
}
\author{
Basile Jumentier
}
