% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2.R
\name{est_oie}
\alias{est_oie}
\title{Overall Indirect Effect for a set of mediator}
\usage{
est_oie(X, m, Y, C, boots = 100)
}
\arguments{
\item{X}{exposure}

\item{m}{a set of mediator}

\item{Y}{outcome}

\item{C}{set of covariable (need to include latent factor)}

\item{boots}{number of bootstrap}
}
\value{
The estimate of Overall Indirect Effect
}
\description{
This function estimate the Overall Indirect Effect
}
\details{
A novelty of HDMAX2 is to evaluate a cumulated indirect effect for all CpGs or AMRs identified in by max2() and .
}
\examples{

#
# res <- est_oie(X, m, Y, C, 10)

}
\author{
Basile Jumentier
}
