% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_OIE.R
\name{estimate_OIE}
\alias{estimate_OIE}
\title{Overall Indirect Effect for a set of mediator}
\usage{
estimate_OIE(X, m, Y, C, boots = 100)
}
\arguments{
\item{X}{exposure}

\item{m}{a set of mediator}

\item{Y}{outcome}

\item{C}{set of covariable (need to include latent factor)}

\item{boots}{number of bootstrap}
}
\value{
The estimate of Overall Indirect Effect
}
\description{
This function estimate the Overall Indirect Effect
}
\details{
A novelty of HDMAX2 is to evaluate a cumulated indirect effect for all mediators
}
\author{
Basile Jumentier
}
