% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2_helper_functions.R
\name{hdmax2_qvalue}
\alias{hdmax2_qvalue}
\title{Compute q-values from p-values}
\usage{
hdmax2_qvalue(pvalues, theta = 0.25)
}
\arguments{
\item{pvalues}{result of max squared test}

\item{theta}{constant defines interquartile range}
}
\value{
qvalues, pi0 =  H0_proportion
}
\description{
Compute q-values from provided p-values Estimate the proportion of H0
relies on the defined interquartile range
}
\examples{
data = hdmax2::simu_data
K = 5
hdmax2_step1 = hdmax2::run_AS(exposure = simu_data$X_binary,
                              outcome = simu_data$Y_continuous,
                              M = simu_data$M1, 
                              K = K)

 #Select candidate mediator  
qv = hdmax2::hdmax2_qvalue(hdmax2_step1$max2_pvalues)



}
\author{
Olivier Francois
}
