% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2_simu_data.R
\docType{data}
\name{helper_ex}
\alias{helper_ex}
\title{Dataset for helper function example}
\format{
a list of 5 objects
\itemize{
\item{methylation}{methylation matrix, 500 individuals and 10000 probes}
\item{exposure}{Exposure for 500 individuals}
\item{phenotype}{Ouctome for 500 individuals}
\item{annotation}{Annotation for the 1000 probes}
\item{covariables}{Covariables for 500 individuals : sex, bmi, age, treatment}
}
}
\usage{
helper_ex
}
\description{
Dataset for helper function example
}
\keyword{datasets}
