% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmm2_med.R
\name{lfmm2_med}
\alias{lfmm2_med}
\title{lfmm2 function}
\usage{
lfmm2_med(input, env, K, lambda = 1e-05, effect.sizes = FALSE)
}
\arguments{
\item{input}{a response variable matrix with n rows and p columns}

\item{env}{An explanatory variable matrix with n rows and d columns.}

\item{K}{latent factor number}

\item{lambda}{ridge penalyzation parameter}

\item{effect.sizes}{true or false to obtain effect sizes}
}
\value{
an object with the following attributes
}
\description{
lfmm2 function
}
\author{
Florence Pittion
}
