% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2_lfmm2_functions.R
\name{lfmm2_med_test}
\alias{lfmm2_med_test}
\title{P-values adjusted for latent factors computed by lfmm2}
\usage{
lfmm2_med_test(
  object,
  input,
  env,
  covar,
  full = FALSE,
  genomic.control = TRUE,
  linear = TRUE,
  family = binomial(link = "logit")
)
}
\arguments{
\item{object}{lfmm2Class object}

\item{input}{a response variable matrix with n rows and p columns}

\item{env}{An explanatory variable matrix with n rows and d columns.}

\item{covar}{covariables}

\item{full}{compute partial regression FALSE/TRUE}

\item{genomic.control}{correct pvalue with genomic inflation factor}

\item{linear}{true or false (else is logistic)}

\item{family}{of logistic reg}
}
\value{
an object with the following attributes
P-values, fscores, zscores, adj.squared, gif
}
\description{
The function returns a vector of p-values for association
between potential mediators and exposure/outcome variables adjusted for latent factors
computed by \code{lfmm2}. As input, it takes an object of class
\code{lfmm2Class} with the data that were used to adjust the LFMM.
If \code{full} is set to \code{FALSE}, the function computes significance
values (p-values) for each exposure variable, otherwise it returns
p-values for the full set of exposure variables.
}
\examples{
attach(simu_data)
K = 5
mod.lfmm1 = lfmm2_med(input = simu_data$M1, 
env = simu_data$X_binary, 
K = K,
effect.sizes = FALSE)

res_reg1 = lfmm2_med_test(mod.lfmm1, 
input = simu_data$M1, 
env = simu_data$X_binary,
covar = cbind(simu_data$age, simu_data$gender),
genomic.control = TRUE)

}
\author{
Florence Pittion, Magali Richard, Olivier Francois
}
