% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmm2_v2.test.R
\name{lfmm2_v2.test}
\alias{lfmm2_v2.test}
\title{lfmm2_test adapted function for hdmax2}
\usage{
lfmm2_v2.test(
  object,
  input,
  env,
  covar = NULL,
  full = FALSE,
  genomic.control = TRUE,
  linear = TRUE,
  family = binomial(link = "logit")
)
}
\arguments{
\item{object}{lfmm2Class object}

\item{input}{a response variable matrix with n rows and p columns}

\item{env}{An explanatory variable matrix with n rows and d columns.}

\item{covar}{covariables}

\item{full}{include covariables true or false}

\item{genomic.control}{correct pvalue with genomic inflation factor}

\item{linear}{true or false (else is logistic)}

\item{family}{of logistic reg}
}
\value{
an object with the following attributes
}
\description{
lfmm2_test adapted function for hdmax2
}
\author{
Florence Pittion
}
