% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2.R
\name{max2}
\alias{max2}
\title{Compute the squared maximum of two series of pValues}
\usage{
max2(pval1, pval2, diagnostic.plot = F, ...)
}
\arguments{
\item{pval1}{vector of p-values of exposure.}

\item{pval2}{vector of p-values of ouctome.}

\item{diagnostic.plot}{if TRUE the histogram of the p-values together
with the estimate of the eta0 null line is plotted.
This is useful to check the fit of the estimated proportion of null p-values visually.}

\item{...}{argument of the fdrtool function from the fdrtool package}
}
\value{
an object with the following attributes:
\itemize{
\item a p-value for each marker
\item a q-value for each marker
\item eta0, an estimate of the proportion of null-hypothesis in all tests
}
}
\description{
This function computes the squared maximum of two series of p-values from the mEWAS() function.
This function tests all potential mediators to evaluate significance the
exposure-outcome association.
}
\details{
A p-value is computed for each marker following the max formula

\deqn{pv = max(pval1, pval2)^2}

The quantity eta0, the proportion of null p-values in a given vector of p-values,
is an important parameter when controlling the false discovery rate (FDR).
A conservative choice is eta0 = 1 but a choice closer to the true value will
increase efficiency and power - see Benjamini and Hochberg (1995, 2000) and Storey (2002) for details.
The function uses the R package fdrtool to transform p-values into q-values and to control the FDR.
}
\examples{

library(hdmax2)

# Run mEWAS

res <- mEWAS(X = example$X, Y = example$Y, M = example$M, K = 5)

# Run max2

res <- max2(pval1 = res$pValue[, 1], pval2 = res$pValue[, 2])

# Manhattan plot

plot(-log10(res$pval), main = paste0("Eta0 = ", round(res$eta0, 3)))
abline(h = -log10(0.05 / ncol(example$M)))

}
\author{
Basile Jumentier
}
