% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2.R
\name{mEWAS}
\alias{mEWAS}
\title{Epigenome Wide Association Study with both exposure and outcome}
\usage{
mEWAS(X, Y, M, K, conf = NULL)
}
\arguments{
\item{X}{(Exposure) a matrix of explanatory variables with n rows and d columns.
Each column corresponds to a distinct explanatory variable.
Explanatory variables must be encoded as numeric variables (matrix).}

\item{Y}{(Outcome) a matrix of explanatory variables with n rows and D columns.
Each column corresponds to a distinct explanatory variable.
Explanatory variables must be encoded as numeric variables.}

\item{M}{a response variable matrix with n rows and p columns.
Each column corresponds to a beta-normalized methylation profile.
Response variables must be encoded as numeric. No NAs are allowed.}

\item{K}{an integer for the number of latent factors in the latent factor
regression model.}

\item{conf}{a matrix with n rows that contains additional covariables (observed confounders), must be numeric.}
}
\value{
an object with the following attributes:
\itemize{
\item U a latent variable score matrix with dimensions n x K.
\item B an effect size matrix for the exposure X and the outcome Y.
\item score a matrix for the exposure X and the outcome Y.
\item pValue a matrix for the exposure X and the outcome Y.
\item calibrated.score2 a calibrated score matrix for the exposure X and the outcome Y.
\item calibrated.pvalue a calibrated p-value matrix for the exposure X and the outcome Y.
\item GIF a Genomic Inflation Factor for exposure and outcome
\item lfmm  the result of the 2 regressions of lfmm, mod1 for the regression of X on M and mod2 for the regression of Y on M given X.
}
}
\description{
This function uses lfmm (latent factor mixed models) to estimate
effect sizes of exposures and outcomes on a response matrix with
beta-normalized methylation data).
}
\details{
The response variable matrix Y and the explanatory variable are centered.
Missing values must be imputed. The number of latent factors can be estimated
by looking at the screeplot of eigenvalues of a PCA.
It allows calibrating scores and p-values by  using the GIF method (Genomic Inflation Factor).
See lfmm package for more information.
}
\examples{

library(hdmax2)

# Run mEWAS

res <- mEWAS(X = example$X, Y = example$Y, M = example$M, K = 5)


}
\author{
Basile Jumentier
}
