% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_estimate_ACME_ADE_PM_TE.R
\name{multi_estimate_ACME_ADE_PM_TE}
\alias{multi_estimate_ACME_ADE_PM_TE}
\title{Estimate effects for a set of mediation markers}
\usage{
multi_estimate_ACME_ADE_PM_TE(
  X_cat,
  Y,
  multi_mediators_top10,
  covar,
  U,
  sims = 3,
  mod2_type,
  ...
)
}
\arguments{
\item{Y}{Outcome. An explanatory variable matrix with n rows and d columns.
Each column corresponds to a distinct explanatory variable (Outcome).
Explanatory variables must be encoded as numeric variables.}

\item{covar}{covariables}

\item{U}{set of latent factors from runAS function (need include covariable)}

\item{sims}{number of Monte Carlo draws for nonparametric bootstrap or quasi-Bayesian approximation.
10000 is recommended.}

\item{mod2_type}{second regression type "linear", "logistic", "surv_Cox"}

\item{...}{argument of the mediate function from the mediation package}

\item{M}{a response variable matrix with n rows and p columns.
Response variables must be encoded as numeric. No NAs allowed.}

\item{X}{Exposure. An explanatory variable matrix with n rows and d columns.
Each column corresponds to a distinct explanatory variable (Exposure).
Explanatory variables must be encoded as numeric variables.}

\item{FDR}{FDR threshold to pass markers in mediation analysis}
}
\value{
Tables of results of mediation analyzes for markers with a Q-value below the FDR threshold.
Composition of tables: estimated effect, confidence interval and mediation pValue.
\itemize{
\item ACME, estimation of the average causal mediation effect (the indirect effect)
\item ADE, estimation average direct effect
\item PM, estimation of the proportion mediated
\item TE, estimation of the total effect
}

Regressions:
\itemize{
\item xm, regression X on M
\item my, regression M on Y
}
}
\description{
Estimate various quantities for causal mediation analysis for a set of
markers, including average causal mediation effects
(indirect effect), average direct effects, proportions mediated,
and total effect.
}
\details{
We use the mediate function of the mediation package on the set of markers having Q-value lower
than the FDR threshold. It estimates their indirect effects and
tests their significance.
}
