% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hdmax2.R
\name{plot_hdmax2}
\alias{plot_hdmax2}
\title{Summary plot for ACME}
\usage{
plot_hdmax2(res_step2, N_med = 10)
}
\arguments{
\item{res_step2}{the table of ACME from the wrap_mediation() function}

\item{N_med}{number of selected mediators}
}
\value{
Summary plot for ACME
}
\description{
This function draw a summary plot of ACME (average causal mediation effect)
}
\examples{

# see wrap_mediation example

}
\author{
Basile Jumentier
}
