% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hdmax_V2.R
\name{plot_V2}
\alias{plot_V2}
\title{Summary plot for HDMAX2
This function draw a summary plot of hdmax2}
\usage{
plot_V2(object, N_med = 10)
}
\arguments{
\item{object}{results from hdmax2 step 2}

\item{N_med}{number of selected mediators}
}
\value{
Summary plot for ACME, PM, OIE, effects interpretation
}
\description{
Summary plot for HDMAX2
This function draw a summary plot of hdmax2
}
\examples{
data(simu_data)
res_step1 = run_AS(X_matrix = simu_data$X_binary ,
Y_matrix = simu_data$Y_continuous,
M_matrix = simu_data$M, 
K = 5,
X_type = "binary",
Y_type = "continuous",
M_type = "methylation",
multivariate = FALSE,
covar = cbind(simu_data$age, simu_data$gender),
diagnostic.plot = F)

mediators_top10 = simu_data$M[,names(sort(res_step1$max2)[1:10])]
res_step2 = estimate_effect(object = res_step1,
m = mediators_top10,
boots = 100,
sims = 100)
plot(res_step2, N_med = 10)
}
\author{
Florence Pittion, Magali Richard
}
