% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.lfmm.R
\name{read.lfmm}
\alias{read.lfmm}
\title{read lfmm}
\usage{
read.lfmm(input.file)
}
\arguments{
\item{input.file}{input file in lfmm object}
}
\value{
an object with the following attributes
}
\description{
read lfmm
}
\author{
Florence Pittion
}
