% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lfmm.R
\name{read_env}
\alias{read_env}
\title{read env}
\usage{
read_env(input.file)
}
\arguments{
\item{input.file}{input file in lfmm object}
}
\value{
an object with the following attributes
}
\description{
read env
}
\author{
Olivier Francois
}
