% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lfmm.R
\name{read_lfmm}
\alias{read_lfmm}
\title{read lfmm}
\usage{
read_lfmm(input.file)
}
\arguments{
\item{input.file}{input file in lfmm object}
}
\value{
an object with the following attributes
}
\description{
read lfmm
}
\author{
Olivier Francois
}
