% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_AS.R
\name{run_AS}
\alias{run_AS}
\title{Association Study with both exposure and outcome}
\usage{
run_AS(
  X_matrix,
  Y_matrix,
  M_matrix,
  K,
  X_type,
  Y_type,
  M_type,
  covar,
  multivariate = FALSE,
  diagnostic.plot = FALSE,
  genomic.control = TRUE,
  effect.sizes = FALSE
)
}
\arguments{
\item{X_matrix}{Exposure. An explanatory variable matrix with n rows and d columns.
Each column corresponds to a distinct explanatory variable (Exposure).
Explanatory variables must be encoded as numeric variables.}

\item{Y_matrix}{Outcome. An explanatory variable matrix with n rows and d columns.
Each column corresponds to a distinct explanatory variable (Outcome).
Explanatory variables must be encoded as numeric variables.}

\item{M_matrix}{a response variable matrix with n rows and p columns.
Response variables must be encoded as numeric. No NAs allowed.}

\item{K}{an integer for the number of latent factors in the regression model.}

\item{X_type}{type of exposition "continuous", "binary", "categorial"}

\item{Y_type}{type of outcome "binary', "continuous"}

\item{M_type}{type of potential mediators matrix "methylation", "transcriptome"}

\item{covar}{set of covariable, must be numeric. No NAs allowed}

\item{multivariate}{choose if for categorial exposure categories treated a single variate or compute in partiel regression (full=TRUE)}

\item{diagnostic.plot}{if TRUE the histogram of the p-values together
with the estimate of eta0 null line is plotted.
Useful to visually check the fit of the estimated proportion of null p-values.}

\item{genomic.control}{correct pvalue with genomic inflation factor}

\item{effect.sizes}{if effect sizes from lfmm are needed}
}
\value{
an object with the following attributes

for first association study (mod1):
\itemize{
\item pValue, estimation of the effects of X and Y on the matrix M.
\item U, scores matrix for the K latent factors, only available for first regression
\item V, latent factors loadings, , only available for first regression
\item zscore, a score matrix for the exposure X and the outcome Y.
\item fscore, a score matrix for the exposure X and the outcome Y.
\item adj_rsquared
\item gif, Genomic inflation factor for X and Y, expressing the deviation of the distribution of the observed test statistic compared to the distribution of the expected test statistic
}

for second association study (mod2):
\itemize{
\item pValue, zscore, fscore,  adj_rsquared, gif
}

results of max2 test:
\itemize{
\item pval, results of max2 test
}

input element:
exposition , outcome, matrix  (element and type) and covar
}
\description{
This function uses lfmm (latent factor mixed models) to estimate
the effects of exposures and outcomes on a response matrix.
applied to estimate the effects of exposure $X$ on a matrix $M$
of potential mediators, and the effect of each marker on outcome $Y$.
It uses the covariables matrix $conf$ and $K$ latent factors.
Test all possible markers to determine potential mediators in the exposure-outcome association.
Then compute the squared maximum of two series of P-values with $max2$ test
It computes the squared maximum of two series of P-values from the association studie.
This rejects the null-hypothesis that either the effect of X on M, or the effect of M on Y is null.
}
\details{
The response variable matrix Y and the explanatory variable are centered.
For each argument, missing values must be imputed: no NA allowed. K (number of latent factors) can be estimated
with the eigenvalues of a PCA.
Possibility of calibrating the scores and pValues by the GIF (Genomic Inflation Factor).
See LEA package for more information.
Max2 test The P-value is computed for each markers following this formula
\deqn{pV = max(pVal1, pVal2)^2}
}
\examples{
data(simu_data)
res_step1 = run_AS(X_matrix = simu_data$X_binary ,
Y_matrix = simu_data$Y_continuous,
M_matrix = simu_data$M, 
K = 5,
X_type = "binary",
Y_type = "continuous",
M_type = "methylation",
multivariate = FALSE,
covar = cbind(simu_data$age, simu_data$gender),
diagnostic.plot = F)
# max2 test results
head(res_step1$max2)
}
\author{
Florence Pittion
}
