% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2_simu_data.R, R/hdmax2data.R
\docType{data}
\name{simu_data}
\alias{simu_data}
\title{Dataset for example}
\format{
A list with  objects
\itemize{
\item  M methylation matrix, 100 samples and 200 potential mediators
\item  M1 methylation matrix, 100 samples and 200 potential mediators
\item  M2 methylation matrix, 100 samples and 200 potential mediators
\item  X_continuous Continuous Exposure for 100 samples
\item  X_continuous2 Continuous Exposure for 100 samples
\item  X_binary Binary Exposure for 100 samples
\item  X_categorial Categorial Exposure for 100 samples
\item  Y_continuous Continuous Outcome for 100 samples
\item  Y_binary Binary Outcome for 100 samples
\item  age age covariable
\item  gender gender covariable
}

A list of ... with ... rows and ... variables
}
\usage{
simu_data

simu_data
}
\description{
A dataset containing blabla. The variables are as follows:
}
\details{
\itemize{
\item X_continuous
\item X_binary, blabla
\item blabla
}
}
\keyword{datasets}
