% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lfmm.R
\name{test_extension}
\alias{test_extension}
\title{test_extension}
\usage{
test_extension(name, extension)
}
\arguments{
\item{name}{input.file}

\item{extension}{extension name}
}
\value{
results
}
\description{
test_extension
}
\author{
Olivier Francois
}
