% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2.R
\name{wrap_mediation}
\alias{wrap_mediation}
\title{Perform mediation analysis for multidimensional mediators (beta-normalized methylation profiles)}
\usage{
wrap_mediation(qval, X, Y, M, U = NULL, FDR = 0.1, sims = 3, ...)
}
\arguments{
\item{qval}{set of q-values obtained from the max2() function}

\item{X}{(Exposure matrix) a matrix of explanatory variables with n rows and d columns.
Each column corresponds to a distinct explanatory variable.
Explanatory variables must be encoded as numeric variables.}

\item{Y}{(Outcome matrix) a matrix of explanatory variables with n rows and D columns.
Each column corresponds to a distinct explanatory variable.
Explanatory variables must be encoded as numeric variables.}

\item{M}{a response matrix with n rows and p columns.
Each column corresponds to a beta-normalized methylation profile.
Response variables must be encoded as numeric. No NAs are allowed.}

\item{U}{set of latent factors from mEWAS() function (needs to include covariable)}

\item{FDR}{FDR threshold in mediation analysis}

\item{sims}{number of Monte Carlo draws for nonparametric bootstrap or quasi-Bayesian approximation.
10000 is recommended.}

\item{...}{argument of the mediate function from the mediation package}
}
\value{
Tables of results of mediation analyzes for markers with a q-value below the FDR threshold.
Indirect effect (ACME - average causal mediation effect), ADE (average direct effect),
PM (proportion mediated) and TE (total effect). Composition of tables: estimated effect,
confidence interval and mediation pValue.

It also returns results of linear regression models.
The xm table corresponds to the regressions of X on M and
the my table to the regressions of Y on M conditional on X.
}
\description{
Estimate various quantities for causal mediation analysis, including average causal mediation effects,
indirect effects, average direct effects, proportion mediated and total effects.
}
\details{
The functions uses the mediate() function of the mediation package on the set of profiles having a q-value lower
than the FDR threshold. This function estimates indirect effects and
tests their significance.
}
\examples{

library(hdmax2)

data(example)

# Run mEWAS

res <- mEWAS(X = example$X, Y = example$Y, M = example$M, K = 5)

# Keep latent factors for mediation

U <- res$U

# Run max2

res <- max2(pval1 = res$pValue[, 1], pval2 = res$pValue[, 2])

# Run mediation (only 3 simulations for estimate and test indirect effect)

res <- wrap_mediation(qval = res$qval,
                            X = example$X,
                            Y = example$Y,
                            M = example$M,
                            U = U, sims = 3,
                            FDR = 0.5)


}
\author{
Basile Jumentier
}
