% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMR_functions.R
\name{wrap_mediation_AMR}
\alias{wrap_mediation_AMR}
\title{Estimate effects for a set of AMR}
\usage{
wrap_mediation_AMR(X, Y, AMR, U = NULL, sims = 3)
}
\arguments{
\item{X}{Exposure. An explanatory variable matrix with n rows and d columns.
Each column corresponds to a distinct explanatory variable (Exposure).
Explanatory variables must be encoded as numeric variables.}

\item{Y}{Outcome. An explanatory variable matrix with n rows and d columns.
Each column corresponds to a distinct explanatory variable (Outcome).
Explanatory variables must be encoded as numeric variables.}

\item{AMR}{a matrix of DMRs from the result AMR_mean of AMR_build function.}

\item{U}{set of latent factors from mEWAS function (need include covariable)}

\item{sims}{number of Monte Carlo draws for nonparametric bootstrap or quasi-Bayesian approximation.
10000 is recommended.}

\item{M}{a response variable matrix with n rows and p columns.
Each column corresponds to a beta-normalized methylation profile.
Response variables must be encoded as numeric. No NAs allowed.}
}
\value{
Tables of results of mediation analyzes for markers with a Q-value below the FDR threshold.
Composition of tables: estimated effect, confidence interval and mediation pValue.
\itemize{
\item ACME, estimation of the average causal mediation effect (the indirect effect)
\item ADE, estimation average direct effect
\item PM, estimation of the proportion mediated
\item TE, estimation of the total effect
}

Regressions:
\itemize{
\item xm, regression X on M
\item my, regression M on Y
}
}
\description{
Estimate various quantities for causal mediation analysis for each
AMRs, including average causal mediation effects
(indirect effect), average direct effects, proportions mediated,
and total effect.
}
\details{
We use the mediate() function of the mediation package on the set of selected AMRs.
This function makes it possible to estimate their indirect effects and to
test their significance.
}
\author{
Basile Jumentier
}
