% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2_helper_functions.R
\name{AMR_build}
\alias{AMR_build}
\title{Build AMR vector}
\usage{
AMR_build(res, methylation, nb_cpg = 2)
}
\arguments{
\item{res}{result object of function AMR_search}

\item{methylation}{a matrix of methylation profile.}

\item{nb_cpg}{threshold of minimal number of CpG in the AMR}
}
\value{
A set of build AMRs.
\itemize{
\item res, selected AMR
\item CpG_for_each_AMR, list of markers present on each AMR.
\item AMR_acp, first components of PCA for each DMR
\item AMR_mean, mean value of CpG on the AMR
}
}
\description{
Build AMR from the result of function AMR_search
}
\details{
We use the series of pValues (one pValue per CpGs) obtained with the mEWAS
regression method and the combination of pValue max2.
To determine the potential AMRs used the combp method present in the ENmix package (Xu et al. 2016).
This method uses the Fisher method to combine the pValues and also the base pair distance (bP)
between CpGs (1000 bP maximum between nb_cpg CpGs on the same AMR).
The information for each AMR is summarized by doing the mean (by row) of each CpG.
}
\examples{
data = hdmax2::helper_ex
#Artificial reduction of dataset size to pass the github action check when building hdmax2 website
data$methylation = data$methylation[ , 800:1000]
data$annotation = data$annotation[800:1000, ]
K=5
## run hdmax2 step1
hdmax2_step1 = hdmax2::run_AS(X_matrix = as.matrix(data$exposure),
                             Y_matrix = as.matrix(data$phenotype),
                             M_matrix = as.matrix(data$methylation),
                             K = K,
                             X_type = "univariate",
                             Y_type = "continuous")

##Detecting AMR
chr = data$annotation$chr
start = data$annotation$start
end = data$annotation$end
pval = hdmax2_step1$max2_pvalues
cpg = data$annotation$cpg

res.amr_search = hdmax2::AMR_search(chr = data$annotation$chr,
                                   start = data$annotation$start,
                                   end = data$annotation$end,
                                   pval = hdmax2_step1$max2_pvalues,
                                   cpg = data$annotation$cpg,
                                   seed = 0.6, #Careful to change this parameter when working with real data
                                   nCores = 2)

res.amr_search$res

res.arm_build = hdmax2::AMR_build(res.amr_search, 
methylation = data$methylation, nb_cpg = 2)
#List of DMR selected
head(res.arm_build$res)
## CpG in the DMR
res.arm_build$CpG_for_each_AMR

}
\author{
Basile Jumentier
}
