% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2.R
\name{AMR_build}
\alias{AMR_build}
\title{Build AMR vector}
\usage{
AMR_build(res, methylation, nb_cpg = 2)
}
\arguments{
\item{res}{result object of AMR_search function}

\item{methylation}{chromosomal position of markers (start)}

\item{nb_cpg}{chromosomal position of markers (end)}
}
\value{
A set of build AMRs.

AMR_mean contains the first components of each PCA for each DMR.
CpG_for_each_AMR contains the list of markers (CpGs) present on each AMR.
}
\description{
To build a vector for each AMR find with AMR_search
}
\details{
We use the series of pValues (one pValue per CpGs) obtained with the mEWAS
regression method and the combination of pValue max2.
To determine the potential AMRs used the combp method present in the ENmix package (Xu et al. 2016).
This method uses the Fisher method to combine the pValues and also the base pair distance (bP)
between CpGs (1000 bP maximum between nb_cpg CpGs on the same AMR).
The information for each AMR is summarized by doing the mean (by row) of each CpG.
}
\examples{

# library(hdma2)
#
# Run mEWAS
#
# res <- mEWAS(X = example$X, Y = example$Y, M = example$M, K = 5)
#
# Keep latent factors for mediation
#
# U <- res$U
#
# Run max2
#
# res <- max2(pval1 = res$pValue[, 1], pval2 = res$pValue[, 2])
#
# lauch AMR_search
#
# res <- AMR_search(chr = example$annotation$chr,
#                   start = example$annotation$start,
#                   end = example$annotation$end,
#                   pval = res$pval,
#                   cpg = example$annotation$cpg, nCores = 1)
#
# lauch AMR_build
#
# tmp <- AMR_build(res, methylation = example$M, nb_cpg = 2)
}
\author{
Basile Jumentier
}
