% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2.R
\name{AMR_search}
\alias{AMR_search}
\title{Find AMR}
\usage{
AMR_search(chr, start, end, pval, cpg, ...)
}
\arguments{
\item{chr}{chromosomes}

\item{start}{chromosomal position of markers (start)}

\item{end}{chromosomal position of markers (end)}

\item{pval}{pValues for each markers, from the max2 function}

\item{cpg}{name of each markers}

\item{...}{see help of combp of ENmix package}
}
\value{
A set of selected AMRs.
}
\description{
To identify aggregated methylated regions (AMR) using a modified comb-p method.
}
\details{
The function will use a modified comb-p method to identify
aggregated methylated regions (AMRs).
}
\examples{

# library(hdma2)
#
# Run mEWAS
#
# res <- mEWAS(X = example$X, Y = example$Y, M = example$M, K = 5)
#
# Keep latent factors for mediation
#
# U <- res$U
#
# Run max2
#
# res <- max2(pval1 = res$pValue[, 1], pval2 = res$pValue[, 2])
#
# lauch AMR_search
#
# res <- AMR_search(chr = example$annotation$chr,
#                   start = example$annotation$start,
#                   end = example$annotation$end,
#                   pval = res$pval,
#                   cpg = example$annotation$cpg, nCores = 1)
#
}
\author{
Basile Jumentier
}
