% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_estimate_effect.R
\name{estimate_effect}
\alias{estimate_effect}
\title{Estimate effects for a set of mediation markers
V2 this version allow to integrate covariates in mediation function}
\usage{
estimate_effect(object, m, boots = 100, ...)
}
\arguments{
\item{object}{results from hdmax2 step 2}

\item{m}{a response variable matrix with n rows and p columns corresponding to mediators selected at step1.
Response variables must be encoded as numeric. No NAs allowed.}

\item{boots}{number of bootstrap}
}
\value{
hdmax2_step2 object
\itemize{
\item ACME, estimation of the average causal mediation effect (the indirect effect)
\item ADE, estimation average direct effect
\item PM, estimation of the proportion mediated
\item TE, estimation of the total effect
}

Regressions:
\itemize{
\item xm, regression X on M
\item my, regression M on Y
}

Overall effect
\itemize{
\item oie, overall indirect effect
\item oie_med , oie median
\item oie_sd , oie standard deviation
\item ote, overall total effect
\item ode, overall direct effect
}
}
\description{
Estimate various quantities for causal mediation analysis for a set of
markers, including average causal mediation effects
(indirect effect), average direct effects, proportions mediated,
and total effect.
}
\details{
We use the mediate function of the mediation package on the set of markers having Q-value lower
than the FDR threshold. It estimates their indirect effects and
tests their significance.
}
\examples{
# Load example dataset
simu_data = hdmax2::simu_data
K = 5
# Run {hdmax2} step 1
hdmax2_step1 = hdmax2::run_AS(
  X = simu_data$X_continuous,
  Y =  simu_data$Y_continuous,
  M =  simu_data$M,
  K = K
)
# Select mediators
mediators_subset = names(sort(hdmax2_step1$max2_pvalues)[1:10])
mediators_top10 = simu_data$M[, mediators_subset]
# Run {hdmax2} step 2
hdmax2_step2 = hdmax2::estimate_effect(object = hdmax2_step1, 
                                       m = mediators_top10)
}
\author{
Florence Pittion, Magali Richard, Olivier Francois, Basile Jumentier
}
