% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2_simu_data.R, R/hdmax2data.R
\docType{data}
\name{helper_ex}
\alias{helper_ex}
\title{Dataset for helper function example}
\format{
a list of 5 objects
\itemize{
\item{methylation}{methylation matrix, 100 individuals and 200 probes}
\item{exposure}{Exposure for 100 individuals}
\item{phenotype}{Ouctome for 100 individuals}
\item{annotation}{Annotation for the 200 probes}
\item{covariables}{Covariables for 100 individuals : sex, bmi, age, treatment}
}

....
}
\usage{
helper_ex

helper_ex
}
\description{
A dataset containing blabla
}
\keyword{datasets}
