% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmm2_med_test.R
\name{lfmm2_med_test}
\alias{lfmm2_med_test}
\title{lfmm2_test adapted function for hdmax2 from LEA::lfmm2.test}
\usage{
lfmm2_med_test(
  object,
  input,
  env,
  covar,
  full = FALSE,
  genomic.control = TRUE,
  linear = TRUE,
  family = binomial(link = "logit")
)
}
\arguments{
\item{object}{lfmm2Class object}

\item{input}{a response variable matrix with n rows and p columns}

\item{env}{An explanatory variable matrix with n rows and d columns.}

\item{covar}{covariables}

\item{full}{compute partial regression FALSE/TRUE}

\item{genomic.control}{correct pvalue with genomic inflation factor}

\item{linear}{true or false (else is logistic)}

\item{family}{of logistic reg}
}
\value{
an object with the following attributes
}
\description{
lfmm2_test adapted function for hdmax2 from LEA::lfmm2.test
}
\examples{
data(simu_data)

mod.lfmm1 = lfmm2_med(input = simu_data$M, 
env = simu_data$X_binary, 
K = 5,
effect.sizes = FALSE)

res_reg1 = lfmm2_med_test(mod.lfmm1, 
input = simu_data$M, 
env = simu_data$X_binary,
covar = cbind(simu_data$age, simu_data$gender),
genomic.control = TRUE)

}
\author{
Florence Pittion
}
