% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2_plot.R
\name{plot_multivariate}
\alias{plot_multivariate}
\title{Summary plot for multivariate exposure variables
This function draw a summary plot of hdmax2}
\usage{
plot_multivariate(object)
}
\arguments{
\item{object}{results from hdmax2 step 2}
}
\value{
Summary plot for ACME, PM, OIE, effects interpretation
}
\description{
Summary plot for multivariate exposure variables
This function draw a summary plot of hdmax2
}
\examples{
# Load example dataset
simu_data = hdmax2::simu_data
X_matrix = simu_data$X_categorial
Y_matrix = simu_data$Y_continuous
M_matrix = simu_data$M
K = 5
# Run {hdmax2} step 1
hdmax2_step1 = hdmax2::run_AS(
  X = X_matrix ,
  Y =  Y_matrix,
  M =  M_matrix,
  K = K
)
# Select mediators
mediators_subset = names(sort(hdmax2_step1$max2_pvalues)[1:10])
mediators_top10 = simu_data$M[, mediators_subset]
# Run {hdmax2} step 2
hdmax2_step2 = hdmax2::estimate_effect(
   object = hdmax2_step1, 
   m = mediators_top10)
hdmax2::plot_multivariate(hdmax2_step2)
}
\author{
Florence Pittion, Magali Richard
}
