% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmax2_plot.R
\name{plot_univariate}
\alias{plot_univariate}
\title{Summary plot for univariate exposure variables
This function draw a summary plot of hdmax2}
\usage{
plot_univariate(object)
}
\arguments{
\item{object}{results from hdmax2 step 2}
}
\value{
Summary plot for ACME, PM, OIE, effects interpretation
}
\description{
Summary plot for univariate exposure variables
This function draw a summary plot of hdmax2
}
\examples{
# Load example dataset
simu_data = hdmax2::simu_data
K = 5
# Run {hdmax2} step 1
hdmax2_step1 = hdmax2::run_AS(
  X = simu_data$X_continuous ,
  Y =  simu_data$Y_continuous,
  M =  simu_data$M,
  K = K
)
# Select mediators
mediators_subset = names(sort(hdmax2_step1$max2_pvalues)[1:10])
mediators_top10 = simu_data$M[, mediators_subset]
# Run {hdmax2} step 2
hdmax2_step2 = hdmax2::estimate_effect(
   object = hdmax2_step1, 
   m = mediators_top10)
hdmax2::plot_univariate(hdmax2_step2)
}
\author{
Florence Pittion, Magali Richard
}
