% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_AS.R
\name{run_AS}
\alias{run_AS}
\title{Association Study with both exposure and outcome}
\usage{
run_AS(
  X_matrix,
  Y_matrix,
  M_matrix,
  K,
  X_type,
  Y_type,
  M_type,
  covar,
  diagnostic.plot = FALSE,
  genomic.control = TRUE,
  effect.sizes = TRUE
)
}
\arguments{
\item{X_matrix}{Exposure. An explanatory variable matrix with n rows and d columns.
Each column corresponds to a distinct explanatory variable (Exposure).
Explanatory variables must be encoded as numeric variables.}

\item{Y_matrix}{Outcome. An explanatory variable matrix with n rows and d columns.
Each column corresponds to a distinct explanatory variable (Outcome).
Explanatory variables must be encoded as numeric variables.}

\item{M_matrix}{a response variable matrix with n rows and p columns.
Response variables must be encoded as numeric. No NAs allowed.}

\item{K}{an integer for the number of latent factors in the regression model.}

\item{X_type}{type of exposition "univariate", "multivariate"}

\item{Y_type}{type of outcome "binary', "continuous"}

\item{M_type}{type of potential mediators matrix "methylation", "transcriptome"}

\item{covar}{set of covariable, must be numeric. No NAs allowed}

\item{diagnostic.plot}{if TRUE the histogram of the p-values together
with the estimate of eta0 null line is plotted.
Useful to visually check the fit of the estimated proportion of null p-values.}

\item{genomic.control}{correct pvalue with genomic inflation factor}
}
\value{
an object with the following attributes
for each association study:
\itemize{
\item U, scores matrix for the K latent factors.
\item V, latent factors loadings
\item effect.sizes , the effect size matrix for the exposure X and the outcome Y.
\item lambda , use in ridge lfmm
\item gif, Genomic inflation factor for X and Y, expressing the deviation of the distribution of the observed test statistic compared to the distribution of the expected test statistic
\item pValue, estimation of the effects of X and Y on the matrix M.
\item zscore, a score matrix for the exposure X and the outcome Y.
\item fscore, a score matrix for the exposure X and the outcome Y.

results of max2 test:
\item pval, results of max2 test
}
}
\description{
This function uses lfmm (latent factor mixed models) to estimate
the effects of exposures and outcomes on a response matrix.
applied to estimate the effects of exposure $X$ on a matrix $M$
of potential mediators, and the effect of each marker on outcome $Y$.
It uses the covariables matrix $conf$ and $K$ latent factors.
Compute the squared maximum of two series of P-values
Test all possible markers to determine potential mediators in the exposure-outcome association.
It computes the squared maximum of two series of P-values from the association studie.
This rejects the null-hypothesis that either the effect of X on M, or the effect of M on Y is null.
}
\details{
The response variable matrix Y and the explanatory variable are centered.
For each argument, missing values must be imputed: no NA allowed. K (number of latent factors) can be estimated
with the eigenvalues of a PCA.
Possibility of calibrating the scores and pValues by the GIF (Genomic Inflation Factor).
See LEA package for more information.
Max2 test The P-value is computed for each markers following this formula

\deqn{pV = max(pVal1, pVal2)^2}

This quantity eta0, i.e. the proportion eta0 of null p-values in a given vector of p-values,
is an important parameter when controlling the false discovery rate (FDR).
A conservative choice is eta0 = 1 but a choice closer to the true value will
increase efficiency and power - see Benjamini and Hochberg (1995, 2000) and Storey (2002) for details.
We use the fdrtool package to transform pValues into qValues,
which allows us to control the FDR.
}
\examples{

data(sample_hdmax2_data)
library(hdmax2)

# Example 1
X_matrix = sample_hdmax2_data$X_binary
Y_matrix = sample_hdmax2_data$Y_time
M_matrix = sample_hdmax2_data$M
age = as.matrix(sample_hdmax2_data$age)
gender = as.matrix(sample_hdmax2_data$gender)
res <- run_AS(X_matrix = X_matrix , Y_matrix = Y_matrix, M_matrix = M_matrix, X_type = "univariate", Y_type = "continuous", K = 5, covar= cbind(age, gender))


}
\author{
Florence Pittion
}
