% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_data.R
\docType{data}
\name{simu_data}
\alias{simu_data}
\title{Dataset for example}
\format{
A list with  objects
\describe{
\item{M}{methylation matrix, 100 samples and 500 potential mediators}
\item{X_continuous}{Continuous Exposure for 100 samples}
\item{X_binary}{Binary Exposure for 100 samples}
\item{X_categorial}{Categorial Exposure for 100 samples}
\item{Y_continuous}{Continuous Outcome for 100 samples}
\item{Y_binary}{Binary Outcome for 100 samples}
\item{age}{age covariable}
\item{gender}{gender covariable}
}
}
\usage{
simu_data
}
\description{
Dataset for example
}
\keyword{datasets}
